:recycle: 시도횟수의 인터페이스는 제거하고 구체 클래스로 대체
